+++ GridboxConnectorAddon-dev/GridboxConnector/utils.py
@@ -2,27 +2,24 @@ import json
 import logging
 import ast
 import os
+import re
 class SensitiveDataFilter(logging.Filter):
     def filter(self, record):
         message = record.getMessage()
         try:
-            literal_msg = ast.literal_eval(message)
+            # Versuche die Nachricht als Python-Dictionary zu parsen
+            message_dict = ast.literal_eval(message)
             # Sensible Daten filtern, falls vorhanden
-            if 'username' in literal_msg:
-                literal_msg['username'] = '***'
-            if 'password' in literal_msg:
-                literal_msg['password'] = '***'
-            if 'id_token' in literal_msg:
-                literal_msg['id_token'] = '***'
-            if 'access_token' in literal_msg:
-                literal_msg['access_token'] = '***'
-            if 'client_id' in literal_msg:
-                literal_msg['client_id'] = '***'
+            sensitive_keys = ['username', 'password', 'id_token', 'access_token', 'client_id', 'token']
+            for key in sensitive_keys:
+                if key in message_dict:
+                    message_dict[key] = '***'
             # Das modifizierte Dictionary zurück in einen String konvertieren
-            record.msg = json.dumps(literal_msg)
+            record.msg = str(message_dict)
+        except (ValueError, SyntaxError):
+            logging.error(f"Error parsing message: {message}")
         except Exception as e:
             logging.error(f"Error filtering sensitive data: {e}")
-            pass
         return True
 
 def get_bool_env(var, default=False):